refactor: Deduplicate GitError and TaskManager imports (Tech Debt 16.2, 16.3)
SADD Workstream Results (ws3):

**GitError Deduplication**:
- Canonical location: core/exceptions.py
- Updated imports in 4 files:
  - aura_cli/doctor.py
  - core/hybrid_loop.py
  - tests/test_git_tools_enhancements.py
  - tests/test_logging_workflow.py
- Backward compatibility preserved via re-exports

**TaskManager Deduplication**:
- Canonical location: core/task_manager.py (confirmed)
- Updated backup file import
- No root task_manager.py existed (already clean)

All tests passing (10/10 affected tests)

Related: Technical Debt 16.2, 16.3
